requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
ptw==1.0.1
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-watch==4.2.0
python-decouple==3.8
setuptools==71.0.1
sqlparse==0.5.1
watchdog==5.0.3
wheel==0.43.0
